Move ContainerImageFromDeployment
